Templates
~Welcome screen~
Clicking on the image takes the user to the login page. Clouds image acts as a proxy. When we decide on a name for our LMS, I can create a banner using Photoshop, maybe.

~Login page~
Login button takes the user to a different site...for now.

~Student view~
Currently working on the design and layout.